move tsdata classes to separate file
